Add function BlogPostController

+++ src/main/java/samuelesimeone/esercizio/exceptions/NotFoundException.java
@@ -1,7 +1,7 @@
 package samuelesimeone.esercizio.exceptions;
 
 public class NotFoundException extends RuntimeException {
-	public NotFoundException(int id) {
+	public NotFoundException(long id) {
 		super("L'utente con id " + id + " non è stato trovato");
 	}
 }
